+++ fitsio/fitsio_pywrap.c
@@ -1066,7 +1066,23 @@ npy_to_fits_table_type(int npy_dtype, int write_bitcols) {
                 PyErr_SetString(PyExc_TypeError, "could not determine 8 byte integer type");
                 return -9999;
             }
-
+#ifdef TULONGLONG
+        case NPY_UINT64:
+            if (sizeof(unsigned long long) == sizeof(npy_uint64)) {
+                return TULONGLONG;
+            } else if (sizeof(unsigned long) == sizeof(npy_uint64)) {
+                return TULONG;
+            } else if (sizeof(unsigned int) == sizeof(npy_uint64)) {
+                return TUINT;
+            } else {
+                PyErr_SetString(PyExc_TypeError, "could not determine 8 byte unsigned integer type");
+                return -9999;
+            }
+#else
+        case NPY_UINT64:
+            PyErr_SetString(PyExc_TypeError, "Unsigned 8 byte integer images are not supported by the FITS standard");
+            return -9999;
+#endif
 
         case NPY_FLOAT32:
             return TFLOAT;
@@ -1080,10 +1096,6 @@ npy_to_fits_table_type(int npy_dtype, int write_bitcols) {
 
         case NPY_STRING:
             return TSTRING;
-
-        case NPY_UINT64:
-            PyErr_SetString(PyExc_TypeError, "Unsigned 8 byte integer images are not supported by the FITS standard");
-            return -9999;
 
         default:
             sprintf(mess,"Unsupported numpy table datatype %d", npy_dtype);
@@ -1178,6 +1190,33 @@ npy_to_fits_image_types(int npy_dtype, int *fits_img_type, int *fits_datatype) {
             }
             break;
 
+#ifdef TULONGLONG
+        case NPY_UINT64:
+            if (sizeof(ULONGLONG) == sizeof(npy_uint64)) {
+                *fits_img_type = ULONGLONG_IMG;
+                *fits_datatype = TULONGLONG;
+            } else if (sizeof(unsigned long) == sizeof(npy_uint64)) {
+                *fits_img_type = ULONG_IMG;
+                *fits_datatype = TULONG;
+            } else if (sizeof(unsigned int) == sizeof(npy_uint64)) {
+                // there is no UINT_IMG, so use ULONG_IMG
+                *fits_img_type = ULONG_IMG;
+                *fits_datatype = TUINT;
+            } else if (sizeof(unsigned long long) == sizeof(npy_uint64)) {
+                // we don't expect to get here
+                *fits_img_type = ULONGLONG_IMG;
+                *fits_datatype = TULONGLONG;
+            } else {
+                PyErr_SetString(PyExc_TypeError, "could not determine 8 byte unsigned integer type");
+                *fits_datatype = -9999;
+                return 1;
+            }
+            break;
+#else
+        case NPY_UINT64:
+            PyErr_SetString(PyExc_TypeError, "Unsigned 8 byte integer images are not supported by the FITS standard");
+            return -9999;
+#endif
 
         case NPY_FLOAT32:
             *fits_img_type = FLOAT_IMG;
@@ -1186,12 +1225,6 @@ npy_to_fits_image_types(int npy_dtype, int *fits_img_type, int *fits_datatype) {
         case NPY_FLOAT64:
             *fits_img_type = DOUBLE_IMG;
             *fits_datatype = TDOUBLE;
-            break;
-
-        case NPY_UINT64:
-            PyErr_SetString(PyExc_TypeError, "Unsigned 8 byte integer images are not supported by the FITS standard");
-            *fits_datatype = -9999;
-            return 1;
             break;
 
         default:
@@ -1293,6 +1326,19 @@ static int fits_to_npy_table_type(int fits_dtype, int* isvariable) {
                 return -9999;
             }
 
+#ifdef TULONGLONG
+        case TULONGLONG:
+            if (sizeof(ULONGLONG) == sizeof(npy_uint64)) {
+                return NPY_UINT64;
+            } else if (sizeof(ULONGLONG) == sizeof(npy_uint32)) {
+                return NPY_UINT32;
+            } else if (sizeof(ULONGLONG) == sizeof(npy_uint16)) {
+                return NPY_UINT16;
+            } else {
+                PyErr_SetString(PyExc_TypeError, "could not determine numpy type for fits TULONGLONG");
+                return -9999;
+            }
+#endif
 
         case TLONGLONG:
             if (sizeof(LONGLONG) == sizeof(npy_int64)) {